% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTaskMasterScript.R
\name{clean_raw_data}
\alias{clean_raw_data}
\title{cleaning and rearranging Raw Data for analysis}
\usage{
clean_raw_data(raw_data)
}
\arguments{
\item{raw_data}{raw dataset that is returned by the \code{read_raw_data}
function}
}
\value{
this function returns the cleaned dataset which is ready
for analysis.
}
\description{
This function takes in the raw dataset and does some basic cleaning (if any)
rearranging (if any).
}
\examples{
clean_raw_data(raw_data = raw_dataset)

}
