% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTaskMasterScript.R
\name{filter_cleaned_data}
\alias{filter_cleaned_data}
\title{filter cleaned data}
\usage{
filter_cleaned_data(cleaned_data, ship_type = "Tanker", ship_name = "MARINUS")
}
\description{
Filters cleaned data using columns \code{ship_type} and \code{ship_name}
}
\details{
@importFrom dplyr filter

@param cleaned_data this is the clean dataset (which is obtained as a
result of cleaning the raw dataset that is read into R.
This dataset will be filtered in this function.

@param ship_type type of ship with the default value "Tanker"
@param ship_name name of ship with the default value "MARINUS"

@return returns cleaned filtered dataset given input values to parameters

@examples
filter_cleaned_data(cleaned_data = cleaned_dataset, ship_type = "Cargo",
ship_name = "MERI")

@export
}
