% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTaskMasterScript.R
\name{max_dist_travelled}
\alias{max_dist_travelled}
\title{Calculate max distance travelled between two (lat, lon) points.}
\usage{
max_dist_travelled(
  cleaned_data,
  filter_ship_type = "Tanker",
  filter_ship_name = "MARINUS"
)
}
\arguments{
\item{cleaned_data}{this is the clean dataset (which is obtained as a
result of cleaning the raw dataset that is read into R}

\item{filter_ship_type}{type of ship with the default value "Tanker"}

\item{filter_ship_name}{name of ship with the default value "MARINUS"}
}
\value{
returns a list containing 2 elements. First element returned is the
\code{max_dist}: max distance. Second element returned is the
\code{max_tibble}: a tibble that contains all data that
correspond to the points b/w which the given ship in question
travelled the maximum distance.
}
\description{
Function that calculates max distance travelled (b/w 2 consecutive points)
from a given ship's travel history. Ship is selected using
\code{filter_ship_type} and \code{filter_ship_name} filters),
}
\examples{
max_dist_travelled(cleaned_data = cleaned_dataset,
filter_ship_type = "Cargo", filter_ship_name = "MERI")


}
