% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTaskMasterScript.R
\name{read_raw_data}
\alias{read_raw_data}
\title{reading raw data into R}
\usage{
read_raw_data(path_to_file)
}
\arguments{
\item{path_to_file}{path to the file where the data is stored}
}
\value{
This function returns an R object which contains the raw dataset.
}
\description{
This functions reads in raw data into R (after performing a few
sanity checks) and throws an error if for any reason data cannot be read.
}
\examples{
read_raw_data(path_to_file = "path/to/data.csv")

}
